pyproject.toml
[tool.poetry]
name = "heo"
version = "0.1.0"
description = ""
authors = ["TheNeoHacker07 <tsstark225@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
pillow = "^10.4.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
redis = "^5.0.7"
celery = "^5.4.0"
django-filter = "^24.2"
pandas = "^2.2.2"
requests-cache = "^1.2.1"
openmeteo-requests = "^1.2.0"
retry-requests = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
